requirements.txt
anyio==3.6.1
asgiref==3.5.2
black==22.6.0
botocore==1.27.41
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
distlib==0.3.5
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.0
fastapi-utils==0.2.1
filelock==3.7.1
flake8==4.0.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
identify==2.5.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
mangum==0.15.0
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.7.0
orjson==3.7.8
pathspec==0.9.0
platformdirs==2.5.2
pre-commit==2.20.0
pycodestyle==2.8.0
pydantic==1.9.1
pyflakes==2.4.0
pynamodb==5.2.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.10
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.16.2
watchgod==0.8.2
websockets==10.3
